third_party/WebKit/LayoutTests/web-animations-api/timed-item-specified-setters.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
var typeError = {name: 'TypeError'};
var timing = new KeyframeEffect(null, null).timing;

test(function() {
    timing.delay = 2;
    assert_equals(timing.delay, 2);
    timing.delay = -4;
    assert_equals(timing.delay, -4);

    assert_throws(typeError, () => timing.delay = NaN);
    assert_equals(timing.delay, -4);

    assert_throws(typeError, () => timing.delay = Infinity);
    assert_equals(timing.delay, -4);
}, 'AnimationEffectTiming should have a setter for delay.');

test(function() {
    timing.endDelay = 0.5;
    assert_equals(timing.endDelay, 0.5);

    timing.endDelay = -5;
    assert_equals(timing.endDelay, -5);

    assert_throws(typeError, () => timing.endDelay = NaN);
    assert_equals(timing.endDelay, -5);

    assert_throws(typeError, () => timing.endDelay = Infinity);
    assert_equals(timing.endDelay, -5);
}, 'AnimationEffectTiming should have a setter for endDelay.');

test(function() {
    timing.fill = 'backwards';
    assert_equals(timing.fill, 'backwards');

    timing.fill = 'both';
    assert_equals(timing.fill, 'both');
}, 'AnimationEffectTiming should have a setter for fill.');

test(function() {
    timing.iterationStart = 1.5;
    assert_equals(timing.iterationStart, 1.5);

    assert_throws(typeError, () => timing.iterationStart = -0.5);
    assert_equals(timing.iterationStart, 1.5);

    assert_throws(typeError, () => timing.iterationStart = NaN);
    assert_equals(timing.iterationStart, 1.5);

    assert_throws(typeError, () => timing.iterationStart = Infinity);
    assert_equals(timing.iterationStart, 1.5);
}, 'AnimationEffectTiming should have a setter for iterationStart.');

test(function() {
    timing.iterations = 10;
    assert_equals(timing.iterations, 10);

    timing.iterations = Infinity;
    assert_equals(timing.iterations, Infinity);

    assert_throws(typeError, () => timing.iterations = -20);
    assert_equals(timing.iterations, Infinity);

    assert_throws(typeError, () => timing.iterations = NaN);
    assert_equals(timing.iterations, Infinity);
}, 'AnimationEffectTiming should have a setter for iterations.');

test(function() {
    timing.duration = Infinity;
    assert_equals(timing.duration, Infinity);

    timing.duration = 1234;
    assert_equals(timing.duration, 1234);

    assert_throws(typeError, () => timing.duration = -10);
    assert_equals(timing.duration, 1234);

    assert_throws(typeError, () => timing.duration = NaN);
    assert_equals(timing.duration, 1234);

    assert_throws(typeError, () => timing.duration = 'very long');
    assert_equals(timing.duration, 1234);
}, 'AnimationEffectTiming should have a setter for duration.');

test(function() {
    timing.playbackRate = 2;
    assert_equals(timing.playbackRate, 2);

    timing.playbackRate = -2;
    assert_equals(timing.playbackRate, -2);

    assert_throws(typeError, () => timing.playbackRate = NaN);
    assert_equals(timing.playbackRate, -2);
}, 'AnimationEffectTiming should have a setter for playbackRate.');

test(function() {
    timing.direction = 'reverse';
    assert_equals(timing.direction, 'reverse');

    timing.direction = 'alternate';
    assert_equals(timing.direction, 'alternate');
}, 'AnimationEffectTiming should have a setter for direction.');

test(function() {
    timing.easing = "step-start";
    assert_equals(timing.easing, 'step-start');

    timing.easing = "eAse\\2d iN-ouT";
    assert_equals(timing.easing, 'ease-in-out');


    assert_throws(typeError, () => timing.easing = 'ponies');
    assert_equals(timing.easing, 'ease-in-out');
}, 'AnimationEffectTiming should have a setter for easing.');
</script>
